+++ packages/gamut/src/DataList/Rows/HeaderRow.tsx
@@ -53,8 +53,8 @@ export const Header: HeaderComponent = ({
         const renderKey = prefixId(`header-col-${rowProperty}`);
         const columnText = header || key;
         return (
-          <ListCol key={renderKey} {...colProps} columnHeader wrap>
-            <FlexBox gap={8} alignItems="flex-end">
+          <ListCol key={renderKey} {...colProps} columnHeader>
+            <FlexBox gap={8} alignItems="flex-end" width="100%">
               {filters && (
                 <FilterControl
                   columnKey={rowProperty}
@@ -63,9 +63,12 @@ export const Header: HeaderComponent = ({
                   justify={colProps.justify}
                 />
               )}
-              {columnText}
-              {sortable && (
-                <SortControl columnKey={rowProperty} onSort={onSort} />
+              {sortable ? (
+                <SortControl columnKey={rowProperty} onSort={onSort}>
+                  {columnText}
+                </SortControl>
+              ) : (
+                columnText
               )}
             </FlexBox>
           </ListCol>